pyproject.toml
[project]
name = "superedudueloge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.1.1",
]
